+++ cmd/cel-key/node_types.go
@@ -9,7 +9,7 @@ import (
 	"github.com/spf13/cobra"
 	flag "github.com/spf13/pflag"
 
-	nodecmd "github.com/celestiaorg/celestia-node/cmd"
+	"github.com/celestiaorg/celestia-node/nodebuilder"
 	"github.com/celestiaorg/celestia-node/nodebuilder/p2p"
 )
 
@@ -55,7 +55,7 @@ func ParseDirectoryFlags(cmd *cobra.Command) error {
 	}
 	switch nodeType {
 	case "bridge", "full", "light":
-		path, err := nodecmd.DefaultNodeStorePath(nodeType, network)
+		path, err := nodebuilder.DefaultNodeStorePath(nodeType, network)
 		if err != nil {
 			return err
 		}
